+++ resources/views/partials/fields/horizontal.blade.php
@@ -1,11 +1,9 @@
-<div class="form-group row">
+<div class="form-group row row-cols-sm-2">
     @isset($title)
         <label for="{{$id}}" class="col-sm-2 text-wrap mt-2 form-label">
             {{$title}}
 
-            @includeWhen(isset($popover),'platform::partials.fields.popover',[
-                'content' => $popover ?? ''
-            ])
+            <x-orchid-popover :content="$popover ?? ''"/>
 
             @if(isset($attributes['required']) && $attributes['required'])
                 <sup class="text-danger">*</sup>
@@ -13,7 +11,7 @@
         </label>
     @endisset
 
-    <div class="col" style="max-width: 440px;">
+    <div class="col">
         {{$slot}}
 
         @if($errors->has($oldName))