+++ tests/Unit/Screen/Fields/RelationTest.php
@@ -5,9 +5,12 @@ declare(strict_types=1);
 namespace Orchid\Tests\Unit\Screen\Fields;
 
 use Illuminate\Database\Eloquent\Collection;
+use Illuminate\Support\Facades\Crypt;
+use Illuminate\Support\Str;
 use Orchid\Platform\Models\Role;
+use Orchid\Platform\Models\User;
 use Orchid\Screen\Fields\Relation;
-use Orchid\Tests\App\AjaxRecord;
+use Orchid\Tests\App\EmptyUserModel;
 use Orchid\Tests\Unit\Screen\TestFieldsUnitCase;
 
 /**
@@ -20,11 +23,17 @@ class RelationTest extends TestFieldsUnitCase
      */
     protected $roles;
 
+    /**
+     * @var Collection|\Illuminate\Database\Eloquent\Model|mixed
+     */
+    protected $users;
+
     public function setUp(): void
     {
         parent::setUp();
 
         $this->roles = Role::factory()->times(10)->create();
+        $this->users = User::factory()->times(10)->create();
     }
 
     public function testInstance(): void
@@ -112,14 +121,67 @@ class RelationTest extends TestFieldsUnitCase
         $this->assertStringContainsString($current[1]->name, $view);
     }
 
-    public function testAJAXClass(): void
+    public function testNotScope(): void
     {
-        $select = Relation::make('role.')
-            ->fromClass(AjaxRecord::class, 'text')
-            ->value(1);
+        $select = Relation::make('users')
+            ->fromModel(EmptyUserModel::class, 'name');
 
         $view = self::renderField($select);
 
-        $this->assertStringContainsString('Record 1', $view);
+        $this->assertStringContainsString('data-fields--relation-scope=""', $view);
+    }
+
+    public function testScopeWithAttributes(): void
+    {
+        $select = Relation::make('users')
+            ->fromModel(EmptyUserModel::class, 'name')
+            ->applyScope('exampleScope', 1, 'string', ['foo', 'bar']);
+
+        $view = self::renderField($select);
+
+        $crypt = Str::between($view, 'data-fields--relation-scope="', '=="');
+
+        $this->assertEquals([
+            'name'       => lcfirst('exampleScope'),
+            'parameters' => [1, 'string', ['foo', 'bar']],
+        ], Crypt::decrypt($crypt));
+    }
+
+    public function testScopeWithNotAttributes(): void
+    {
+        $select = Relation::make('users')
+            ->fromClass(EmptyUserModel::class, 'name')
+            ->applyScope('exampleScope');
+
+        $view = self::renderField($select);
+
+        $crypt = Str::between($view, 'data-fields--relation-scope="', '=="');
+
+        $this->assertEquals([
+            'name'       => lcfirst('exampleScope'),
+            'parameters' => [],
+        ], Crypt::decrypt($crypt));
+    }
+
+    public function testScopeWithInstance(): void
+    {
+        /** @var User $current */
+        $current = $this->users->random();
+
+        // With parameters
+        $select = Relation::make('users')
+            ->value($current->id)
+            ->fromClass(EmptyUserModel::class, 'name')
+            ->applyScope('asFilerId', $current->id);
+
+        $this->assertStringContainsString($current->name, self::renderField($select));
+
+        // Invalid parameter
+        $select = Relation::make('users')
+            ->value($current->id)
+            ->fromClass(EmptyUserModel::class, 'name')
+            ->applyScope('asFilerId', 0);
+
+        $this->assertStringNotContainsString($current->name, self::renderField($select));
     }
 }